thoughtflow/requirements.txt
Django==6.0.3
djangorestframework==3.16.1
django-cors-headers==4.9.0
Pillow==11.3.0
